Move some nonstd to a private header
